requirements.txt
alembic==1.13.1
asttokens==2.4.1
biopython==1.83
blinker==1.7.0
click==8.1.7
colorama==0.4.6
colorlog==6.8.0
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.0
fonttools==4.47.2
greenlet==3.0.3
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lightgbm==4.2.0
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.9
numpy==1.26.3
optuna==3.5.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
wcwidth==0.2.13
Werkzeug==3.0.1
xgboost==2.0.3
